Added basic AI to the computer player
Player will now look at all possible move and then select the highest
scoring one.

Score is determined by the weight of the bone (add the 2 numbers) and
then a bias added for doubles since they are harder bones (stop
sniggering Tim) to get rid of.

Double bias is currently set to 5. The scoring bit is in the possible
plays constructor.