Major graphics rework
Merged Abbie's FramePanels into Artist. Then split out a couple of
derived types of JPanels.

Abbie - the key bit was overriding the JPanel's paint method with a
version that added the bits we needed and then calls the parent class'
version.

Next thing to do is to sort out my two additional classes -
PlayerHandPanel and TablePanel. There is a load of repeated code so they
should be derived from a shared parent class.

Note all my reworks have killed the text output we had. Can always put
it back in - preferably as a class separate to the graphics ones.